Cast the multiband GeoTIFFs, and fetch the merged layer
Both found by watching the Drive task list after a run that reported success.

1. The feature rasters never arrived

    Exported bands must have compatible data types; found inconsistent
    types: Float64 and Int32. (Error code: 3)

Both multiband rasters bolt an integer `cluster_id` band onto float feature
bands. A comment in export.py asserted that "GeoTIFF casts the whole image to a
common type, so it rides along as a float next to the feature bands". It does
not; EE rejects the export.

The failure mode is what makes this worth a test rather than a one-line fix.
The rejection is asynchronous: the stage submits the task, logs a task id,
reports success, and the pipeline finishes clean. Three seconds later the task
fails on a page nobody was watching, and the file is simply absent from Drive.
Nothing in the run says so.

Both images are now cast with `.toFloat()`. Float32 rather than Float64: it
holds cluster ids 0..k-1 exactly, carries ~7 significant digits across every
feature band (elevation in metres, radar in dB, indices in [-1, 1], the scaled
stack in roughly [-5, 5]), and halves the file. The single-band uint8 label
raster stays integer and is the authoritative copy of the labels; cluster_id
rides along in the two multiband rasters for convenience.

2. The fetcher could not see the layer just added to the export

`fetch_drive_vectors.py` defaulted to `("dissolved", "snic")`, so it would have
downloaded two layers, reported no error, and left any comparison running on the
pre-merge ones -- with `stands_merged` sitting untouched in Drive. Default is now
`("merged", "dissolved", "snic")` in both the function signature and the CLI,
which were two separate defaults for one decision.

Testing: 5 new tests. The raster ones read the export block and assert each
multiband spec is cast while the uint8 label raster is not -- a live check would
need the async task result, which is exactly what nobody saw. The fetcher ones
pin merged in the default set, that suffix matching does not confuse the three
layer names, and that the CLI default still matches the signature's. 503 pass,
ruff clean on CI's scope, no new mypy errors.

+++ src/fmu/stages/export.py
@@ -221,9 +221,25 @@ class ExportStage(Stage):
         #                             k-means saw it (log / robust-scaled,
         #                             cyclic-decomposed) + the cluster_id band.
         if config.export.export_geotiff:
-            # cluster_id band, shared by the two feature rasters. Kept as a
-            # plain band (GeoTIFF casts the whole image to a common type, so
-            # it rides along as a float next to the feature bands).
+            # cluster_id band, shared by the two feature rasters.
+            #
+            # The whole image is cast to Float32 before export. GeoTIFF does
+            # NOT cast a mixed-type image to a common type -- the previous
+            # comment here claimed it did, and EE rejects it instead:
+            #
+            #   Exported bands must have compatible data types; found
+            #   inconsistent types: Float64 and Int32. (Error code: 3)
+            #
+            # And it rejects it *asynchronously*, three seconds into a task
+            # submitted long after the stage reported success, so the run looked
+            # clean and the file simply never arrived.
+            #
+            # Float32 rather than Float64: it holds cluster ids 0..k-1 exactly,
+            # carries ~7 significant digits across every feature band (elevation
+            # in metres, radar in dB, indices in [-1, 1], the scaled stack in
+            # roughly [-5, 5]), and halves the file. The authoritative integer
+            # labels are the single-band uint8 raster below; cluster_id rides
+            # along in these two for convenience.
             label_band = cluster_labels.rename("cluster_id")
 
             raster_specs = [
@@ -235,12 +251,12 @@ class ExportStage(Stage):
                 (
                     "raster_features_raw",
                     f"{config.name}_features_raw",
-                    _build_raw_feature_export_image(ctx).addBands(label_band),
+                    _build_raw_feature_export_image(ctx).addBands(label_band).toFloat(),
                 ),
                 (
                     "raster_features_scaled",
                     f"{config.name}_features_scaled",
-                    ctx.get("feature_stack").addBands(label_band),
+                    ctx.get("feature_stack").addBands(label_band).toFloat(),
                 ),
             ]
             for manifest_key, raster_filename, raster_image in raster_specs:
